coastalme/runner/coastalme-runner/requirements.txt
pyqt5~=5.15.4
numpy==1.26.1
pandas==2.1.2
python-dateutil~=2.8.2
seaborn==0.13.0